Gestion des fichiers spéciaux

+++ client/include/gestionner.h
@@ -17,7 +17,7 @@ typedef struct REP_MSG *REP_MSG;
 
 struct REP_MSG{
     int state;
-    int size;
+    long long size;
 };
 
 int server_gestionner(int clientfd,rio_t rio, char* buf);
